refactor: PieceDao 생성, GameDao에서 Piece 테이블 종속적인 메소드 삭제
